Extract settings error popup helper
