+++ src/helpers/settingsPage.js
@@ -1,25 +1,7 @@
 import { loadSettings, updateSetting } from "./settingsUtils.js";
 import { fetchDataWithErrorHandling } from "./dataUtils.js";
 import { DATA_DIR } from "./constants.js";
-
-/**
- * Display a temporary popup when a settings update fails.
- *
- * @pseudocode
- * 1. Remove any existing popup element with the `.settings-error-popup` class.
- * 2. Create a new div with that class containing an error message.
- * 3. Append the div to `document.body`.
- * 4. Remove the popup after 2 seconds.
- */
-export function showSettingsError() {
-  const existing = document.querySelector(".settings-error-popup");
-  existing?.remove();
-  const popup = document.createElement("div");
-  popup.className = "settings-error-popup";
-  popup.textContent = "Failed to update settings.";
-  document.body.appendChild(popup);
-  setTimeout(() => popup.remove(), 2000);
-}
+import { showSettingsError } from "./showSettingsError.js";
 
 function applyInputState(element, value) {
   if (!element) return;
